Changed format of project structure, added endpoints, fixed others.
Queries are now updated to search for non-deleted information, file structure has been changed for easier readability and organization.
Endpoints are available to get lists of sessions from game ids and campaign ids.
