+++ src/statego-backend/.vscode/settings.json
@@ -1,5 +1,7 @@
 {
     "rust-analyzer.linkedProjects": [
+        ".\\Cargo.toml",
+        ".\\Cargo.toml",
         ".\\Cargo.toml"
     ]
 }
