+++ src/statego-backend/src/main.rs
@@ -27,20 +27,28 @@ async fn main() -> io::Result<()> {
         // serve functions at defined endpoints and bind global data pool
         App::new()
             .app_data(shared_data.clone())
-            .service(routes::create_user)
-            .service(routes::get_users)
-            .service(routes::hello)
-            .service(routes::login)
-            .service(routes::update_user_profile)
-            .service(routes::create_session)
-            .service(routes::get_list_of_sessions)
-            .service(routes::create_game)
-            .service(routes::create_campaign)
-            .service(routes::get_list_of_games)
-            .service(routes::get_list_of_campaigns)
-            .service(routes::get_single_user)
-            .service(routes::delete_session)
-            .service(routes::get_single_session)
+            .service(routes::user::create_user)
+            .service(routes::user::get_users)
+            .service(routes::user::hello)
+            .service(routes::user::login)
+            .service(routes::user::update_user_profile)
+            .service(routes::user::get_single_user)
+            .service(routes::session::create_session)
+            .service(routes::session::get_list_of_sessions)
+            .service(routes::session::delete_session)
+            .service(routes::session::get_list_of_sessions_from_campaign)
+            .service(routes::session::get_list_of_sessions_from_game)
+            .service(routes::session::get_single_session)
+            .service(routes::game::create_game)
+            .service(routes::game::get_single_game)
+            .service(routes::game::get_list_of_games)
+            .service(routes::campaign::create_campaign)
+            .service(routes::campaign::get_single_campaign)
+            .service(routes::campaign::get_list_of_campaigns)
+            .service(routes::campaign::delete_campaign)
+            
+            
+            
     })
     .bind(("127.0.0.1", 8080))?
     .workers(2)
