+++ src/statego-backend/src/models/mod.rs
@@ -1,0 +1,4 @@
+pub mod campaign;
+pub mod game;
+pub mod session;
+pub mod user;
