+++ src/statego-backend/src/persistence/mod.rs
@@ -1,0 +1,6 @@
+pub mod campaign;
+pub mod error;
+pub mod game;
+pub mod session;
+pub mod tools;
+pub mod user;
