--- src/statego-backend/src/queries.rs
@@ -1,485 +1,0 @@
-/////////////////////////////////////////////
-use crate::models::*;
-/// queeries.rs
-///
-/// handles querying the mysql database
-/////////////////////////////////////////////
-use mysql::{params, prelude::*};
-
-pub fn insert_new_ueser(
-    conn: &mut mysql::PooledConn,
-    my_email: String,
-    my_username: String,
-    my_password: String,
-    my_first_name: String,
-    my_last_name: String,
-) -> mysql::error::Result<u64> {
-    conn.exec_drop(
-        "
-        INSERT INTO users (email, username, pass, first_name, last_name)
-        VALUES (:email, :username,:pass, :first_name, :last_name)
-        ",
-        params! {
-            "email" => my_email,
-            "username" => my_username,
-            "pass" => my_password,
-            "first_name" => my_first_name,
-            "last_name" => my_last_name,
-        },
-    )
-    .map(|_| conn.last_insert_id())
-}
-
-//TODO select all fields from the user table
-pub fn select_all_users(conn: &mut mysql::PooledConn) -> mysql::error::Result<Vec<UserData>> {
-    conn.query_map(
-        "
-        SELECT id, email, username, first_name, last_name
-        FROM users
-        ",
-        |(my_id, my_email, my_username, my_first_name, my_last_name)| UserData {
-            id: my_id,
-            email: my_email,
-            username: my_username,
-            first_name: my_first_name,
-            last_name: my_last_name,
-        },
-    )
-}
-
-pub fn select_password_by_username(
-    conn: &mut mysql::PooledConn,
-    username: String,
-) -> mysql::error::Result<String> {
-    conn.exec_first(
-        "
-        SELECT pass
-        FROM users
-        WHERE username = :username
-        ",
-        params! {
-            "username" => username
-        },
-    )
-    .map(|pass| pass.unwrap())
-}
-
-pub fn select_single_user(
-    conn: &mut mysql::PooledConn,
-    id: u64,
-) -> mysql::error::Result<SingleUserUnconvertedResponseData> {
-    conn.exec_first(
-        "
-        SELECT email, username, first_name, last_name, pronouns, bio, profile_pic
-        FROM users
-        WHERE id = :id
-        ",
-        params! {
-            "id" => id
-        }, //|(email, username, first_name, last_name, pronouns, bio, profile_pic)| SingleUserUnconvertedResponseData {
-           //    email: email,
-           //    username: username,
-           //    first_name: first_name,
-           //    last_name: last_name,
-           //    pronouns: pronouns,
-           //    bio: bio,
-           //    profile_pic: profile_pic
-           //}
-    )
-    .map(|single_user_response| single_user_response.unwrap())
-}
-
-pub fn select_user_by_id(
-    conn: &mut mysql::PooledConn,
-    username: String,
-) -> mysql::error::Result<UserData> {
-    conn.exec_first(
-        "
-        SELECT id, email, username, first_name, last_name
-        FROM users
-        WHERE username = :username
-        ",
-        params! {
-            "username" => username
-        },
-    )
-    .map(|user| user.unwrap())
-}
-
-pub fn update_bio_and_profilepic(
-    conn: &mut mysql::PooledConn,
-    username: String,
-    bio: String,
-    profile_pic: String,
-) -> mysql::error::Result<UserUpdateData> {
-    conn.exec_drop(
-        r"UPDATE users 
-        SET bio = :bio, profile_pic = :profile_pic 
-        WHERE username = :username",
-        params! {
-            "bio" => bio,
-            "profile_pic" => profile_pic,
-            "username" => username.clone()
-        },
-    );
-    conn.exec_first(
-        "
-        SELECT username, bio, profile_pic
-        FROM users
-        WHERE username = :username
-        ",
-        params! {
-            "username" => username
-        },
-    )
-    .map(|user_update| user_update.unwrap())
-}
-
-pub fn create_session_in_database(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-    game_id: u64,
-    campaign_id: Option<u64>,
-    session_start_string: String,
-    session_end_string: String,
-    players: String,
-    number_of_players: i8,
-    notes: Option<String>,
-    winner: bool,
-    winner_name: Option<String>,
-    picture: Option<String>,
-) -> mysql::error::Result<u64> {
-    conn.exec_drop(
-        "
-        INSERT INTO session (user_id, game_id, campaign_id, 
-        session_start, session_end, players, notes, winner, winner_name, picture, number_of_players)
-        VALUES (:user_id, :game_id, :campaign_id, 
-            :session_start, :session_end, :players, :notes, :winner, :winner_name, :picture, :number_of_players)
-        ",
-        params! {
-            "user_id" => user_id,
-            "game_id" => game_id,
-            "campaign_id" => campaign_id,
-            "session_start" => session_start_string,
-            "session_end" => session_end_string,
-            "players" => players,
-            "notes" => notes,
-            "winner" => winner,
-            "winner_name" => winner_name,
-            "picture" => picture,
-            "number_of_players" => number_of_players
-        },
-    )
-    .map(|_| conn.last_insert_id())
-}
-
-pub fn delete_session_in_database(
-    conn: &mut mysql::PooledConn,
-    session_id: u64,
-    is_deleted: bool
-) -> mysql::error::Result<u64> {
-        conn.exec_drop(
-            r"UPDATE session 
-            SET is_deleted = :is_deleted
-            WHERE session_id = :session_id",
-            params! {
-                "session_id" => session_id,
-                "is_deleted" => is_deleted
-            }
-        )
-        .map(|_| conn.affected_rows())
-}
-
-pub fn select_gameid_by_gamestring(
-    conn: &mut mysql::PooledConn,
-    game_name: String,
-) -> mysql::error::Result<u64> {
-    conn.exec_first(
-        "
-        SELECT game_id
-        FROM games
-        WHERE game_name = :game_name
-        ",
-        params! {
-            "game_name" => game_name
-        },
-    )
-    .map(|game_id| game_id.unwrap())
-}
-
-pub fn select_gamestring_by_gameid(
-    conn: &mut mysql::PooledConn,
-    game_id: u64,
-) -> mysql::error::Result<String> {
-    conn.exec_first(
-        "
-        SELECT game_name
-        FROM games
-        WHERE game_id = :game_id
-        ",
-        params! {
-            "game_id" => game_id
-        },
-    )
-    .map(|game_name| game_name.unwrap())
-}
-
-pub fn select_campaignid_by_campaignstring(
-    conn: &mut mysql::PooledConn,
-    campaign_name: Option<String>,
-) -> mysql::error::Result<u64> {
-    conn.exec_first(
-        "
-        SELECT campaign_id
-        FROM campaign
-        WHERE campaign_name = :campaign_name
-        ",
-        params! {
-            "campaign_name" => campaign_name
-        },
-    )
-    .map(|campaign_id| campaign_id.unwrap())
-}
-
-pub fn select_campaignstring_by_campaignid(
-    conn: &mut mysql::PooledConn,
-    campaign_id: u64,
-) -> mysql::error::Result<String> {
-    conn.exec_first(
-        "
-        SELECT campaign_name
-        FROM campaign
-        WHERE campaign_id = :campaign_id
-        ",
-        params! {
-            "campaign_id" => campaign_id
-        },
-    )
-    .map(|campaign_id| campaign_id.unwrap())
-}
-pub fn select_userid_by_userstring(
-    conn: &mut mysql::PooledConn,
-    username: String,
-) -> mysql::error::Result<u64> {
-    conn.exec_first(
-        "
-        SELECT id
-        FROM users
-        WHERE username = :username
-        ",
-        params! {
-            "username" => username
-        },
-    )
-    .map(|user_id| user_id.unwrap())
-}
-
-pub fn select_userstring_by_userid(
-    conn: &mut mysql::PooledConn,
-    id: u64,
-) -> mysql::error::Result<String> {
-    conn.exec_first(
-        "
-        SELECT username
-        FROM users
-        WHERE id = :id
-        ",
-        params! {
-            "id" => id
-        },
-    )
-    .map(|username| username.unwrap())
-}
-
-
-
-pub fn get_single_session_query(
-    conn: &mut mysql::PooledConn,
-    session_id: u64,
-) -> mysql::error::Result<Vec<SessionDataUnConverted>> {
-    conn.exec_map(
-        "
-        SELECT session_id, user_id, game_id, campaign_id, session_start, session_end, players, notes, winner, winner_name, picture, number_of_players
-        FROM session
-        WHERE session_id = :session_id
-        ",
-        params! {
-            "session_id" => session_id
-        }
-        ,
-        |(session_id, user_id, game_id, campaign_id, session_start, session_end, 
-            players, notes, winner, winner_name, picture, 
-            number_of_players)| SessionDataUnConverted {
-            session_id : session_id,
-            user_id : user_id,
-            game_id : game_id,
-            campaign_id : campaign_id,
-            session_start: session_start,
-            session_end: session_end,
-            players: players,
-            notes: notes,
-            winner: winner,
-            winner_name: winner_name,
-            picture: picture,
-            number_of_players: number_of_players
-        }
-    )
-}
-
-pub fn get_list_of_sessions_queries(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-    game_id: u64,
-    campaign_id: Option<u64>,
-) -> mysql::error::Result<Vec<SessionDataUnConverted>> {
-    if campaign_id.is_none() {
-        conn.exec_map(
-            "
-            SELECT session_id, user_id, game_id, campaign_id, session_start, session_end, players, notes, winner, winner_name, picture, number_of_players
-            FROM session
-            WHERE user_id = :user_id AND game_id = :game_id AND campaign_id IS NULL
-            ",
-            params! {
-                "user_id" => user_id,
-                "game_id" => game_id,
-            "campaign_id" => campaign_id
-            }
-            ,
-
-            |(session_id, user_id, game_id, campaign_id, session_start, session_end, 
-                players, notes, winner, winner_name, picture, 
-                number_of_players)| SessionDataUnConverted {
-                session_id : session_id,
-                user_id : user_id,
-                game_id : game_id,
-                campaign_id : campaign_id,
-                session_start: session_start,
-                session_end: session_end,
-                players: players,
-                notes: notes,
-                winner: winner,
-                winner_name: winner_name,
-                picture: picture,
-                number_of_players: number_of_players
-            }
-        )
-    } else {
-        conn.exec_map(
-            "
-            SELECT session_id, user_id, game_id, campaign_id, session_start, session_end, players, notes, winner, winner_name, picture, number_of_players
-            FROM session
-            WHERE user_id = :user_id AND game_id = :game_id AND campaign_id = :campaign_id
-            ",
-            params! {
-                "user_id" => user_id,
-                "game_id" => game_id,
-            "campaign_id" => campaign_id
-            }
-            ,
-
-            |(session_id, user_id, game_id, campaign_id, session_start, session_end, 
-                players, notes, winner, winner_name, picture, 
-                number_of_players)| SessionDataUnConverted {
-                session_id : session_id,
-                user_id : user_id,
-                game_id : game_id,
-                campaign_id : campaign_id,
-                session_start: session_start,
-                session_end: session_end,
-                players: players,
-                notes: notes,
-                winner: winner,
-                winner_name: winner_name,
-                picture: picture,
-                number_of_players: number_of_players
-            }
-        )
-    }
-}
-
-pub fn create_game_in_database(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-    game_name: String,
-    descr: Option<String>,
-) -> mysql::error::Result<u64> {
-    conn.exec_drop(
-        "
-        INSERT INTO games (user_id, game_name, descr)
-        VALUES (:user_id, :game_name, :descr)
-        ",
-        params! {
-            "user_id" => user_id,
-            "game_name" => game_name,
-            "descr" => descr,
-
-        },
-    )
-    .map(|_| conn.last_insert_id())
-}
-
-pub fn create_campaign_in_database(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-    game_id: u64,
-    campaign_name: String,
-    descr: Option<String>,
-    notes: Option<String>,
-) -> mysql::error::Result<u64> {
-    conn.exec_drop(
-        "
-        INSERT INTO campaign (user_id, game_id, campaign_name, descr, notes)
-        VALUES (:user_id, :game_id, :campaign_name, :descr, :notes)
-        ",
-        params! {
-            "user_id" => user_id,
-            "game_id" => game_id,
-            "campaign_name" => campaign_name,
-            "descr" => descr,
-            "notes" => notes
-
-        },
-    )
-    .map(|_| conn.last_insert_id())
-}
-
-pub fn get_list_of_games_queries(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-) -> mysql::error::Result<Vec<GameInfo>> {
-    conn.exec_map(
-        "
-        SELECT game_name, descr FROM games WHERE user_id = :user_id
-        ",
-        params! {
-            "user_id" => user_id
-        },
-        |(game_name, descr)| GameInfo {
-            game_title: game_name,
-            description: descr,
-        },
-    )
-}
-
-pub fn get_list_of_campaigns_queries(
-    conn: &mut mysql::PooledConn,
-    user_id: u64,
-    game_id: u64,
-) -> mysql::error::Result<Vec<CampaignInfo>> {
-    conn.exec_map(
-        "
-        SELECT campaign_name, descr, notes FROM campaign WHERE user_id = :user_id AND game_id = :game_id
-        ",
-        params! {
-            "user_id" => user_id,
-            "game_id" => game_id
-        }
-        ,
-
-        |( campaign_name, descr, notes)| CampaignInfo {
-            campaign_title: campaign_name,
-            description: descr,
-            notes: notes
-        },
-    )
-}
