+++ src/statego-backend/src/queries/mod.rs
@@ -1,0 +1,4 @@
+pub mod campaign;
+pub mod game;
+pub mod session;
+pub mod user;
