+++ src/statego-backend/src/routes/mod.rs
@@ -1,0 +1,4 @@
+pub mod campaign;
+pub mod game;
+pub mod session;
+pub mod user;
